Add alternate array-based format, add clarification to match option
The `match` option for relationships now must be `false` explicitly to
force create a relationship. Otherwise it will only create the
relationship if one of the same type and direction do not already exist.

The neo4j.py script was removed in favor of the graphlib implementation
noted in the README.